Remove BiblicalBytes projects from solution
This commit removes the `BiblicalBytes`, `BiblicalBytes.Desktop`, `BiblicalBytes.Bible.Presentation`, and `BiblicalBytes.Bible.Module` projects from the `Biblical-Bytes.sln` solution. It includes the deletion of Debug and Release configurations for these projects and the removal of their nesting within the solution's folders. Additionally, the reference to the `BiblicalBytes` project has been removed from the `BiblicalBytes.Desktop.csproj` file, indicating that it no longer depends on or links to the `BiblicalBytes` project.